Handle correctly servers not initiating close
If we use the default handler to a Redis server, Switcher will keep the
connection open forever, because Redis server is not going to initiate
the close of the connection.

With this new scheme, as soon as one channel has finished sending data,
the two connections are closed and Shovel could finish.

This follows the pattern seen in https://github.com/google/tcpproxy/blob/master/tcpproxy.go#L386